requirements.txt
pandas>=2.2.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
PyYAML>=6.0.1
watchdog>=3.0.0
python-dotenv>=1.0.0